+++ config/routes.rb
@@ -85,6 +85,7 @@ Rails.application.routes.draw do
   get "booking/reschedule/:appointment_hash" => "booking#reschedule"
   post "booking/get_available_hours" => "booking#get_available_hours"
   get "booking/get_unavailable_dates" => "booking#get_unavailable_dates"
+  post "booking/confirm" => "booking#confirm"
   post "booking/register" => "booking#register"
   get "booking_confirmation/of/:appointment_hash" => "booking_confirmation#of", as: :booking_confirmation
   get "booking_confirmation/ics/:appointment_hash" => "booking_confirmation#ics", as: :booking_confirmation_ics